divide personInfo by new Component, decomposes logic. Add universal component
