+++ src/components/elements/user-profile-card/UserProfileCard.tsx
@@ -22,175 +22,17 @@ import { FormProvider, useForm } from "react-hook-form";
 import VerifyUserData from "@/components/ui/verify-user-data/VerifyUserData";
 import ChangeUserDataForm from "@/components/ui/change-user-data-form/ChangeUserDataForm";
 import ProfileField from "@/components/ui/profile-field/ProfileField";
+import { UserInfo } from "@/types/state.interface";
 
-// Интерфейсы для состояния поля и модального окна
-export interface FieldState {
-  id: string;
-  name: string;
-  label?: string;
-  placeholder?: string;
-  type: string;
-  componentType: ComponentFormEnum;
+interface IUserProfileCardProp {
+  userInfo: UserInfo;
+  openModal: (...args: any[]) => void;
 }
 
-interface ModalState {
-  isOpen: boolean;
-  title: string;
-  buttonText: string;
-  field: FieldState;
-}
-
-const UserProfileCard: React.FC = () => {
-  const dispatch = useAppDispatch();
-  const userInfo = useAppSelector((state) => state.USER);
-  const userInfoFull = { ...userInfo, organization: "ООО Велесъ" };
-
-  useEffect(() => {
-    async function fetchUserInfo() {
-      try {
-        // const response: UserServerData = await UserService.getUserInfo();
-        // const userInfo = adaptToUserData(response);
-        // console.log(userInfo);
-
-        const user = {
-          id: 1, // Уникальный идентификатор пользователя
-          email: "john.doe@example.com", // Email пользователя
-          fullName: "John Doe", // Полное имя пользователя
-          birthDate: "1990-01-01", // Дата рождения пользователя
-          phone: "+1234567890", // Номер телефона пользователя
-          isActive: true, // Статус активности пользователя
-          isSuperuser: true, // Статус суперпользователя
-          isVerified: false, // Статус верификации пользователя
-          picture: "https://example.com/picture.jpg", // Ссылка на фото пользователя
-          isAuth: "AUTHENTICATED", // Статус аутентификации пользователя
-          organization: "ООО Велесъ", // Статус аутентификации пользователя
-        };
-
-        // Обновляет данные пользователя в сторе
-        dispatch(setUserInfo(user));
-
-        // dispatch(setUserInfo(userInfo));
-      } catch (error) {
-        console.error("Error fetching user info:", error.message);
-      }
-    }
-
-    fetchUserInfo();
-  }, [dispatch]);
-
-  // Верификация почты
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  async function verifyUserEmail() {
-    try {
-      // Иммитация запроса
-      setTimeout(() => {
-        dispatch(setUserInfo({ isVerified: true }));
-      }, 2000);
-
-      // setIsVerified(true);
-
-      // TODO:
-      // const response = await UserService.verifyUserEmail();
-
-      // if (response.status == 200) {
-      //   console.log("User created successfully:", response.data);
-      // }
-    } catch (error) {
-      console.error("Failed to verified user:", error.response);
-    }
-  }
-
-  const handleVerifiedEmail = () => {
-    setIsModalOpen(true);
-    verifyUserEmail();
-  };
-
-  // Изменить Личные данные
-  const [modalState, setModalState] = useState<ModalState>({
-    isOpen: false,
-    title: "",
-    buttonText: "",
-    field: {
-      id: "",
-      name: "",
-      label: "",
-      placeholder: "",
-      type: "",
-      componentType: ComponentFormEnum.INPUT,
-    },
-  });
-
-  const openModal = (
-    title: string,
-    buttonText: string,
-    field: Partial<FieldState>
-  ) => {
-    setModalState({
-      isOpen: true,
-      title,
-      buttonText,
-      field: {
-        id: field.id || "",
-        name: field.name || "",
-        label: field.label || "",
-        placeholder: field.placeholder || "",
-        type: field.type || "",
-        componentType: field.componentType || ComponentFormEnum.INPUT,
-      },
-    });
-  };
-
-  const closeModal = () => {
-    setModalState({
-      isOpen: false,
-      title: "",
-      buttonText: "",
-      field: {
-        id: "",
-        name: "",
-        label: "",
-        placeholder: "",
-        type: "",
-        componentType: ComponentFormEnum.INPUT,
-      },
-    });
-  };
-
-  const handleModalSave = (data: any) => {
-    console.log(data);
-
-    dispatch(
-      setUserInfo({
-        ...userInfo,
-        [modalState.field.name]: data[modalState.field.name],
-      })
-    );
-
-    closeModal();
-    methods.reset({});
-  };
-
-  console.log(userInfo);
-
-  // Hook форма.
-  const methods = useForm({
-    mode: "onChange",
-    defaultValues: {},
-  });
-
-  const text = (
-    <>
-      Для завершения регистрации и доступа ко всем функциям нашего сервиса,
-      пожалуйста, подтвердите вашу электронную почту. На указанный адрес:
-      <b>&nbsp;{userInfo.email}&nbsp;</b> было отправлено письмо с инструкциями
-      по подтверждению.
-      <br />
-      Если вы не получили письмо, проверьте папку "Спам" или нажмите кнопку
-      ниже, чтобы отправить письмо повторно.
-    </>
-  );
-
+const UserProfileCard: React.FC<IUserProfileCardProp> = ({
+  userInfo,
+  openModal,
+}) => {
   return (
     <>
       <div className="">
@@ -214,6 +56,7 @@ const UserProfileCard: React.FC = () => {
                     name: "fullName",
                     placeholder: "Введите полное имя",
                     type: "text",
+                    componentType: ComponentFormEnum.INPUT,
                   })
                 }
               />
@@ -227,6 +70,7 @@ const UserProfileCard: React.FC = () => {
                     name: "email",
                     placeholder: "Введите почту",
                     type: "text",
+                    componentType: ComponentFormEnum.INPUT,
                   })
                 }
               />
@@ -240,6 +84,7 @@ const UserProfileCard: React.FC = () => {
                     name: "phone",
                     placeholder: "Введите телефон",
                     type: "text",
+                    componentType: ComponentFormEnum.INPUT,
                   })
                 }
               />
@@ -253,13 +98,14 @@ const UserProfileCard: React.FC = () => {
                     name: "birthDate",
                     placeholder: "Введите дату рождения",
                     type: "date",
+                    componentType: ComponentFormEnum.INPUT,
                   })
                 }
               />
 
               <ProfileField
                 label="Организация"
-                value={userInfoFull.organization}
+                value={userInfo.organization}
                 // onEditClick={() => {}}
                 isEditable={false}
               />
@@ -274,51 +120,8 @@ const UserProfileCard: React.FC = () => {
           ) : null}
         </div>
 
-        <div className="mt-8 border border-zinc-200 rounded p-3">
-          <div className="flex items-center justify-between">
-            <p className="text-gray-500">
-              Подтверидите вашу почту: &nbsp;
-              <span className="text-black">{userInfo.email}</span>
-            </p>
-
-            {userInfo.isVerified ? (
-              <CircleCheck size={24} className="stroke-green-500" />
-            ) : (
-              <Button
-                className="bg-blue-500 text-white"
-                onClick={handleVerifiedEmail}
-              >
-                Подтвердить
-              </Button>
-            )}
-          </div>
-        </div>
+      
       </div>
-
-      {/* Подтверждение почты */}
-      {isModalOpen && !userInfo.isVerified ? (
-        <Modal className="w-1/2 max-w-xl" onClose={() => setIsModalOpen(false)}>
-          <VerifyUserData
-            title="Подтвердите вашу почту"
-            text={text}
-            buttonText="Отправить письмо повторно"
-            handleButtonClick={handleVerifiedEmail}
-          />
-        </Modal>
-      ) : null}
-
-      {/* Редактировать данные */}
-      {modalState.isOpen ? (
-        <Modal className="w-1/2 max-w-xl" onClose={closeModal}>
-          <ChangeUserDataForm
-            title={modalState.title}
-            methods={methods}
-            field={modalState.field}
-            buttonText={modalState.buttonText}
-            handleFormSave={methods.handleSubmit(handleModalSave)}
-          />
-        </Modal>
-      ) : null}
     </>
   );
 };
